pyproject.toml
[project]
name = "langchain-openlit-demo"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5",
    "openlit>=1.33.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.4.0",
    "python-dotenv>=1.0.0",
]
